Add locale folder and place the pages files into so the translation works.
