+++ src/app/[locale]/pokemon-guides/page.tsx
